Setting layout for home page and top level pages
The layout for the home page will display the most recent blog post. Top
level pages such as an about me page will display a similar format as
the home page, but the top level content will not display information
about the post date or the title of the post.