Implementing categories taxonomy
The categories are tied by whatever is saved in a post and listed on the
sidebar of the webpage. Therefore, if we have 5 posts and each post is
categorized under Go, then that category will be listed with a count of
5. Categories are also referenced on each post and will be listed on the
footer of the post when on the list post view. Showing them on the
content post view may be considered later.

A new site params variable has been created called categoriesByCount.
This will be used to order the categories Alphabetically or ByCount.
Setting the variable to true will order the categories ByCount.
Note: if you do not create the variable in your config then the
categories will be ordered Alphabetically.

Create three new $.Scratch variables called customTitle, categories, and
showCategories.
    customTitle:    Will be used to set a custom title for the webpage
                    before calling the partial template, header

    categories:     Used to save the OrderedTaxonomy which is used to
                    loop through the elements of our categories and list
                    the count per category

    showCategories: a boolean value used in the partial template,
                    sidebar to display the categories or not. Setting
                    the value to false will hide the categories. Not
                    setting a value will show the list