Implement sharing feature and set post limit
Sharing buttons will show up on each post. The same buttons will show
appear on the home page since the page displays the content of the most
recent post. The same sharing buttons will also appear on the navigation
bar, but only if the page type is a post and if the page is the home
page.

Introducing new $.Scratch variable called shareNav which will show the
share icon on the navigation bar when the value is equal to true. This
only appears from what was mentioned in the previous paragraph.

Creating new Site variable called Params.socialAppearAtTop and
Params.socialAppearAtBottom. These two variables are used in the sidebar
and will show the social icons if set to true. The icons will not appear
if the variables are not created in the config file.

The most recent posts amount on the sidebar can be dynamically altered
with the new Site variable called Params.postAmount.sidebar. If the
variable is not created in the config file then a default amount of 5
will be used.