Adding ability to load minified css and js files
The two new site param variables are called:
    minifiedFilesCSS and minifiedFilesJS

The variables will be an array that will allow the user to input
multiple min files if needed. The min should consist of the base files
provided in this theme. The user then has the ability to add their own
modifications if needed.