Can now add alt text to an image
The alt text value can be provided in the intro pic by setting the site
param:
    [params.intro.pic]
        alt = "set text"

The shortcodes img-fit and img-post can now set the alt text value.
The parameter for img-post is the last position or the named parameter
alt. Here is an example for both:
    {{< img-post "date" "file.jpg" "left" "alt text" >}}
    {{< img-post path="date" file="file.jpg" type="left" alt="text" >}}

img-fit has changed slightly. The first parameter will not be the path.
Instead, that will always be the last parameter. You still have the
ability to place as many images as needed in this shortcode call.
    postion:
        0: width of column 1-12 in u or u$
        1: filename including extension
        2: alt text
        3: same as pos 0
        4: same as pos 1
        5: same as pos 2
        ...
        n-3
        n-2
        n-1
        n: name of the path