Removing pre tag css and adding new shortcode
The custom css that wraps the pre tag text has been removed. Instead,
make sure the line of code does not exceed 80 characters. If it does,
then continue on the next line. The reason why is because exceeding
the limit conflicts with the css in the overall look of the website.

Creating new shortcode called url-link which is a positional 2 or 3
parameter shortcode.
    position 0: Name of the link
    position 1: URL of the link
    position 2: (optional) the target of the URL. Default value is
                _blank