Adding CSS fix for pre code and markdown images
The pre code is no longer constrained by 80 characters per line.
Exceeding the limit no longer conflicts with the overall look of the
website. This resolves the following issue:
https://github.com/jpescador/hugo-future-imperfect/issues/8

Images have also been fixed if the user decides to utilize the standard
markdown functionality instead of the custom image shortcodes