Fix category url in category-footer.html
Ensure that there is a trailing '/' in the category URL when using the
url defined for the menu item "Categories". Without this the template
produces invalid urls for the category links in the footer of the post
if the user has defined a menu item named "Categories" without a
trailing slash (as in the example page).