Change references to resources to use relURL
I have updated all the links to js, css, img, and other files that are
local to use the relURL command in the template. This allows for base
URLs that are nested below the root (ie. https://mysite.com/blog instead of
just https://blog.mysite.com/ ). This commit fixes #22.